pyproject.toml
[tool.poetry]
name = "central_database"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytz = "2022.1"  # https://github.com/stub42/pytz
python-slugify = "6.1.2"  # https://github.com/un33k/python-slugify
Pillow = "9.1.1"  # https://github.com/python-pillow/Pillow
argon2-cffi = "21.3.0"  # https://github.com/hynek/argon2_cffi
whitenoise = "6.2.0"  # https://github.com/evansd/whitenoise
redis = "4.3.3"  # https://github.com/redis/redis-py
hiredis = "2.0.0"  # https://github.com/redis/hiredis-py
uvicorn = {extras = ["standard"], version = "0.17.6"}  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django = "3.2.14"  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ = "0.9.0" # https://github.com/joke2k/django-environ
django-model-utils = "4.2.0"  # https://github.com/jazzband/django-model-utils
django-allauth = "0.51.0"  # https://github.com/pennersr/django-allauth
django-crispy-forms = "1.14.0"  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5 = "0.6"  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis = "5.2.0"  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework = "3.13.1"  # https://github.com/encode/django-rest-framework
django-cors-headers = "3.13.0" # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular = "0.22.1"  # https://github.com/tfranzel/drf-spectacular
django-anymail = {extras = ["mailgun"], version = "8.6"}  # https://github.com/anymail/django-anymail

gunicorn = "20.1.0"  # https://github.com/benoitc/gunicorn
psycopg2 = "2.9.3"  # https://github.com/psycopg/psycopg2
sentry-sdk = "1.6.0"  # https://github.com/getsentry/sentry-python
dj-rest-auth = {extras = ["with-social"], version = "^2.2.5"}

[tool.poetry.dev-dependencies]
Werkzeug = {extras = ["watchdog"], version = "2.0.3"} # https://github.com/pallets/werkzeug
ipdb = "0.13.9"  # https://github.com/gotcha/ipdb
psycopg2 = "2.9.3"  # https://github.com/psycopg/psycopg2
watchgod = "0.8.2"  # https://github.com/samuelcolvin/watchgod

# Testing
# ------------------------------------------------------------------------------
mypy = "0.950"  # https://github.com/python/mypy
django-stubs = "1.12.0"  # https://github.com/typeddjango/django-stubs
pytest = "7.1.2"  # https://github.com/pytest-dev/pytest
pytest-sugar = "0.9.4"  # https://github.com/Frozenball/pytest-sugar
djangorestframework-stubs = "1.7.0"  # https://github.com/typeddjango/djangorestframework-stubs
pytest-cov = "3.0.0" # https://github.com/pytest-dev/pytest-cov/

# Documentation
# ------------------------------------------------------------------------------
sphinx = "5.0.1"  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild = "2021.3.14" # https://github.com/GaretJax/sphinx-autobuild
sphinxcontrib-httpdomain = "1.8.0" # https://github.com/sphinx-contrib/httpdomain
sphinxcontrib-openapi = "0.7.0" # https://github.com/sphinx-contrib/openapi
mistune = "0.8.4" # https://github.com/lepture/mistune

# Code quality
# ------------------------------------------------------------------------------
flake8 = "4.0.1"  # https://github.com/PyCQA/flake8
flake8-isort = "4.1.1"  # https://github.com/gforcada/flake8-isort
coverage = "6.4.1"  # https://github.com/nedbat/coveragepy
black = "22.6.0"  # https://github.com/psf/black
pylint-django = "2.5.3"  # https://github.com/PyCQA/pylint-django
pre-commit = "2.19.0"  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy = "3.2.1"  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar = "3.5.0" # https://github.com/jazzband/django-debug-toolbar
django-extensions = "3.1.5"  # https://github.com/django-extensions/django-extensions
django-coverage-plugin = "2.0.3"  # https://github.com/nedbat/django_coverage_plugin
pytest-django = "4.5.2"  # https://github.com/pytest-dev/pytest-django
pydotplus = "2.0.2"
graphviz = "^0.20.1"

[build-system]
requires = ["poetry==1.1.14"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "central_database.config.settings"
